api/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
requests==2.31.0
Pillow==9.5.0
beautifulsoup4==4.12.2
lxml==4.9.3